User aggegators. Fixed errors and mispellings

+++ lab_14_user_aggregations/V14.1.1__create_complex_number_type.sql
@@ -1,4 +1,4 @@
 CREATE TYPE public.complex_number AS(
-	real DOUBLE,
-	imaginary DOUBLE
+	real NUMERIC,
+	imaginary NUMERIC
 );
